% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgrtelomeres-package.R
\docType{package}
\name{cgrtelomeres-package}
\alias{cgrtelomeres}
\alias{cgrtelomeres-package}
\title{cgrtelomeres: Automated Processing of Telomere qPCR Data}
\description{
Process CGR lab telomere qPCR output into standard
    report formats.
}
\details{

}
\author{
\strong{Maintainer}: Cameron Palmer \email{cameron.palmer@nih.gov}

}
