% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_analysis.R
\name{create.analysis}
\alias{create.analysis}
\title{Create a PrimaryAnalysis from a loaded ExportDatum}
\usage{
create.analysis(
  export.datum,
  plate.content.report = NA,
  plate.list = NA,
  infer.384.locations = FALSE
)
}
\arguments{
\item{export.datum}{An ExportDatum object with raw qPCR data for processing}

\item{plate.content.report}{Character vector of plate content report for
project (e.g. "PlateContentReport_GP0317-TL1.xls") or NA}

\item{plate.list}{Character vector of plate list for project (e.g.
"PlateList_GP0317-TL1.xls") or NA}

\item{infer.384.locations}{Logical: whether to assume fixed 96->384
well mapping}
}
\value{
a PrimaryAnalysis object containing the processed results
for the input
}
\description{
\code{create.analysis} takes an ExportDatum object containing telomeric
and control data and applies the standard analysis currently being run
by the excel template that generates files in Data/Analysis/*xlsx.
}
\details{
This function attempts to emulate the results in the primary analysis
spreadsheet. There's a fair amount of redundant information (log2 transform
of some data, linear regression information that is deprecated) that
is not being included for the moment.
}
\examples{
filename.pair <- find.input.files("Examples for Bioinformatics")
export.datum <- read.export.datum(filename.pair)
processed.analysis <- create.analysis(export.datum)
}
\seealso{
\code{\link[=PrimaryAnalysis]{PrimaryAnalysis()}} for constructing empty PrimaryAnalysis objects;
\code{\link[=read.export.datum]{read.export.datum()}} for generating compatible input data structures.
}
\keyword{telomeres}
