% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{create.output.directories}
\alias{create.output.directories}
\title{Create output directory structure if necessary}
\usage{
create.output.directories(top.path)
}
\arguments{
\item{top.path}{Character vector of output top-most directory}
}
\description{
\code{create.output.directories} creates all directories and subdirectories
required for the standard output of the telomere protocol. If the top-level
directory does not exist, it will also be created.
}
\details{
The output directories are: top/Data/Analysis, top/Analysis/QC,
top/Analysis/Final. This silently handles if said folders already exist.
}
\examples{
create.output.directories("C:/Users/palmercd/Documents/telomeres/results")
}
\keyword{telomeres}
