% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_datum.R
\docType{class}
\name{ExportDatum-class}
\alias{ExportDatum-class}
\title{Store aligned telomere and control data per experiment}
\description{
Store aligned telomere and control data per experiment
}
\section{Slots}{

\describe{
\item{\code{Analysis.Code}}{character vector describing alphabetic analysis
code (A-H)}

\item{\code{Cp.Telo}}{numeric vector of cycle threshold data for telomere data}

\item{\code{Cp.Control}}{numeric vector of cycle threshold data for control gene
data}

\item{\code{Standard}}{numeric vector of standard concentration, or NA if not a
standard sample}

\item{\code{Well.ID}}{character vector of well ID codes from Source Plate Contents}

\item{\code{Vial.ID}}{character vector of vial ID codes from Source Plate Contents}
}}

\examples{
new("ExportDatum",
  Cp.Telo = runif(3, 0, 38),
  Cp.Control = runif(3, 0, 38),
  Standard = rep(NA, 3),
  Well.ID = rep(NA, 3),
  Vial.ID = rep(NA, 3)
)
}
\seealso{
\code{\link[=read.export.datum]{read.export.datum()}} for loading data into \code{ExportDatum}
}
\keyword{telomeres}
