% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_analysis.R
\name{extract.well.mappings}
\alias{extract.well.mappings}
\title{Map 96 to 384 wells using Plate List and Content Report data}
\usage{
extract.well.mappings(plate.list.data, plate.content.data, analysis.code)
}
\arguments{
\item{plate.list.data}{Data frame of plate list data}

\item{plate.content.data}{Data frame of plate content report data}

\item{analysis.code}{Character vector: case-sensitive letter \link{A-H} for run code}
}
\value{
List with entries Rep1.Well, Rep2.Well, Rep3.Well, corresponding to
columns in the Data/Analysis/*xlsx files. Each will be a factor.
}
\description{
\code{extract.well.mappings} takes plate list and content data and extracts
the actual 96->384 well mapping used in the analysis, as opposed to
the default one observed in the initial test case.
}
\details{
The output is sorted lexicographically by 96 well label. Replicate
assignments (1 vs 2 vs 3) are similarly resolved by lexicographical
sort of the 384 well labels. This should compel the output to exactly
match the previous protocol's Data/Analysis/*xlsx format if used
without further shuffling.
}
\examples{
content.filename <- "PlateContentReport_GP0317-TL1.xls"
list.filename <- "PlateList_GP0317-TL1.xls"
content.data <- openxlsx:read.xlsx(content.filename, sheet = 1, rowNames = FALSE, colNames = TRUE)
list.data <- openxlsx::read.xlsx(list.filename, sheet = 1, rowNames = FALSE, colNames = TRUE)
analysis.code <- "H"
well.mappings <- extract.well.mappings(list.data, content.data, analysis.code)

}
\seealso{
\code{\link[=create.analysis]{create.analysis()}} for calling function; \code{\link[openxlsx:read.xlsx]{openxlsx::read.xlsx()}} for
an example of how to generate the plate list or content data frames.
}
\keyword{telomeres}
