% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{format.final.analysis}
\alias{format.final.analysis}
\title{Make final report spreadsheet}
\usage{
\method{format}{final.analysis}(primary.analysis, project.id)
}
\arguments{
\item{primary.analysis}{PrimaryAnalysis object to be manipulated}

\item{project.id}{Character vector of project ID (e.g. "GP0317-TL7")}
}
\description{
\code{format.final.analysis} compels the contents of a PrimaryAnalysis
object into the appropriate data.frame format and returns it, expecting
it to be combined with the results from other PrimaryAnalysis objects
with \code{rbind}.
}
\details{
Exact format details are TBD.
}
\seealso{
\code{\link[=create.analysis]{create.analysis()}} for generating compatible
PrimaryAnalysis objects.
}
\keyword{telomeres}
