% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{hello.world}
\alias{hello.world}
\title{Test function for gitlab communication}
\usage{
hello.world(ignored)
}
\description{
\code{hello.world} prints said text to screen.
}
\details{
This is entirely a placeholder and will be removed.
}
