% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_analysis.R
\name{process.average}
\alias{process.average}
\title{Compute replicate mean with restrictions}
\usage{
process.average(vec.rep1, vec.rep2, vec.rep3)
}
\arguments{
\item{vec.rep1}{numeric vector of first replicate Ct data}

\item{vec.rep2}{numeric vector of second replicate Ct data}

\item{vec.rep3}{numeric vector of third replicate Ct data}
}
\value{
numeric vector of vector-wise averages, with NA
exclusion criteria applied
}
\description{
\code{process.average} takes three replicates' data and combines vector-wise
into means while allowing a single observation from each condition at
most to be missing (flagged as NA).
}
\details{
There is an implicit restriction that appears to be applied manually to the
real data example, where conditions with fewer than two replicates are
blanked. Hopefully, this should cover that situation, while allowing
single dropouts.
}
\examples{
vec.mean <- process.average(rnorm(10), rnorm(10), rnorm(10))
}
\seealso{
\code{\link[=process.standard.deviation]{process.standard.deviation()}} for equivalent calculation
with standard deviation.
}
