% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{process.experiment}
\alias{process.experiment}
\title{Process a CGR-style telomere qPCR experiment}
\usage{
process.experiment(input.path, output.path, project.id)
}
\arguments{
\item{input.path}{Character vector of input directory path}

\item{output.path}{Character vector of output directory path}

\item{project.id}{Character vector of project ID (e.g. "GP0317-TL7")}
}
\description{
\code{process.experiment} takes the output of a lab telomere qPCR experiment
and generates the expected output content from prior processing methods.
}
\details{
This is the main entry point for the CGR telomere processing package.
The input directory is sanity checked for its expected contents. The output
directory will be created if it doesn't already exist. Any existing conflicting
output files will be overwritten by this method, so please change output.path
if you don't want this behavior.
}
\examples{
process.experiment("Examples for Bioinformatics", "C:/Users/palmercd/Documents/telomeres/output", "GP0317-TL7")

}
\keyword{telomeres}
