% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_datum.R
\name{read.export.datum}
\alias{read.export.datum}
\title{Read and align Cp data for telomeres and controls in an experiment}
\usage{
read.export.datum(exp.control.filenames)
}
\arguments{
\item{exp.control.filenames}{length-2 vector of filenames: one for Telo, one for 36B4}
}
\value{
an instance of S4 class \code{ExportDatum} containing aligned relevant information.
}
\description{
\code{read.export.datum} reads a pair of exported files containing separated telomere and control
data and aligns them for later use
}
\details{
This function removes extraneous metadata and provides mild control checking. It does
not postprocess the data in any meaningful way. This function should not be called
directly by the end user.
}
\examples{
read.export.datum(c("Data/Exports/PC29625_A_Telo.txt", "Data/Exports/PC29624_A_36B4.txt"))

}
\seealso{
\code{\link[=find.input.files]{find.input.files()}} for generating the expected input for this function.
}
\keyword{telomeres}
