% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{report.primary.analysis}
\alias{report.primary.analysis}
\title{Write PrimaryAnalysis object to file in simplified format}
\usage{
report.primary.analysis(primary.analysis, output.path, project.id)
}
\arguments{
\item{primary.analysis}{PrimaryAnalysis object to be written to file}

\item{output.path}{Character vector of output directory path}

\item{project.id}{Character vector of project ID (e.g. "GP0317-TL7")}
}
\description{
\code{report.primary.analysis} takes a processed PrimaryAnalysis object
and reports it to file in a format analogous to the report format
from the existing excel template.
}
\details{
Exact format details are TBD.
}
\examples{
my.files <- find.input.files("Data/Exports")
my.data <- read.export.datum(my.files[[1]])
my.analysis <- create.analysis(my.data)
report.primary.analysis(my.analysis, "my.output", "GP0317-TL7")

}
\seealso{
\code{\link[=create.analysis]{create.analysis()}} for generating compatible PrimaryAnalysis objects.
}
\keyword{telomeres}
