% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgrtelomeres-package.R
\docType{package}
\name{cgrtelomeres-package}
\alias{cgrtelomeres}
\alias{cgrtelomeres-package}
\title{cgrtelomeres: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
