% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_datum.R
\docType{class}
\name{ExportDatum-class}
\alias{ExportDatum-class}
\title{Store aligned telomere and control data per experiment}
\description{
Store aligned telomere and control data per experiment
}
\section{Slots}{

\describe{
\item{\code{Pos}}{factor of well locations for datapoints on a 384 well plate}

\item{\code{Cp.Telo}}{numeric vector of cycle threshold data for telomere data}

\item{\code{Cp.Control}}{numeric vector of cycle threshold data for control gene data}

\item{\code{Standard}}{numeric vector of standard concentration, or NA if not a standard sample}
}}

\examples{
new("ExportDatum", Pos = factor(c("A1", "A2", "A3")),
                   Cp.Telo = runif(3, 0, 38),
                   Cp.Control = runif(3, 0, 38),
                   Standard = rep(NA, 3))

}
\seealso{
\code{\link[=read.export.datum]{read.export.datum()}} for loading data into \code{ExportDatum}
}
\keyword{telomeres}
