% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{process.experiment}
\alias{process.experiment}
\title{Process a CGR-style telomere qPCR experiment}
\usage{
process.experiment(
  input.path,
  output.path,
  project.id = NA,
  plate.content.report = NA,
  plate.list = NA,
  infer.384.locations = FALSE,
  subject.list.from.input.path = TRUE,
  markdown.html = NA
)
}
\arguments{
\item{input.path}{Character vector of input directory path}

\item{output.path}{Character vector of output directory path}

\item{project.id}{Character vector of project ID
(e.g. "GP0317-TL7") or NA}

\item{plate.content.report}{Character vector of plate content
report for project (e.g. "PlateContentReport_GP0317-TL1.xls") or NA}

\item{plate.list}{Character vector of plate list for project
(e.g. "PlateList_GP0317-TL1.xls") or NA}

\item{infer.384.locations}{Logical: whether to assume fixed
96->384 well mapping}

\item{subject.list.from.input.path}{Logical: whether to pull subject
data from Data/Analysis/*xlsx of an existing analysis run}

\item{markdown.html}{character vector: if specified, this will be
the name of the output report generated by the software; otherwise
specify NA to suppress markdown output}
}
\value{
TBD
}
\description{
\code{process.experiment} takes the output of a lab telomere qPCR experiment
and generates the expected output content from prior processing methods.
}
\details{
This is the main entry point for the CGR telomere processing package.
The input directory is sanity checked for its expected contents. The
output directory will be created if it doesn't already exist. Any
existing conflicting output files will be overwritten by this method,
so please change output.path if you don't want this behavior.

During development, plate content report and list files were not
necessarily available for test cases. As such, an override mode
has been added that pulls pre-pasted copies of relevant content
report data from the spreadsheets in Data/Analysis from an existing
telomere run. During deployment, this will almost certainly not be
the desired behavior, and this feature is scheduled to be removed.
This flag will also be overridden when plate list and content report
are both specified.

In all existing test datasets, a fixed mapping from 96 well to 384 well
layouts has been found. For mild efficiency purposes, this mapping
has been set as the default. If there is ever any doubt, or if the
well mapping is changed at some point in the future, setting
\code{infer.384.locations} to \code{FALSE} will trigger dynamic remapping using
the data in \code{plate.content.report} and \code{plate.list}.
}
\examples{
process.experiment(
  "Examples for Bioinformatics",
  "C:/Users/palmercd/Documents/telomeres/output", "GP0317-TL7"
)
}
\keyword{telomeres}
